refactor: use the new function to DRY the code
after relizing we would have to repeat the code to parse the proc/stat file,
I created a function for the parsing that just
takes the keyword.
the user of the funciton will need to know
the index of the value being sought, and
we assume the first word in the line is the keyword.